+++ src/core/map/sprite/being.cpp
@@ -510,7 +510,7 @@ void Being::drawSpeech(const int offsetX, const int offsetY)
 {
     const int px = mPx - offsetX;
     const int py = mPy - offsetY;
-    const int speech = (int) config.getValue("speech", NAME_IN_BUBBLE);
+    const int speech = config.getValue("speech", NAME_IN_BUBBLE);
 
     // Draw speech above this being
     if (mSpeechTime > 0 && (speech == NAME_IN_BUBBLE ||
